Remove unnecessary processor base class
